Add compound operators and Shift operator
