+++ crates/bollard-ls/src/operator/scramble.rs
@@ -18,8 +18,3 @@
 // LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-pub mod compound;
-pub mod local_search_operator;
-pub mod shift;
-pub mod swap;
